+++ netpyne/cell.py
@@ -6,7 +6,9 @@ Contains Cell related classes
 Contributors: salvadordura@gmail.com
 """
 
+from numbers import Number
 from neuron import h # Import NEURON
+from specs import Dict
 import sim
 
 
@@ -22,8 +24,8 @@ class Cell (object):
     def __init__ (self, gid, tags, create=True, associateGid=True):
         self.gid = gid  # global cell id 
         self.tags = tags  # dictionary of cell tags/attributes 
-        self.secs = {}  # dict of sections
-        self.secLists = {}  # dict of sectionLists
+        self.secs = Dict()  # dict of sections
+        self.secLists = Dict()  # dict of sectionLists
         self.conns = []  # list of connections
         self.stims = []  # list of stimuli
 
@@ -55,10 +57,14 @@ class Cell (object):
                 if condVal not in self.tags['label']:
                     conditionsMet = 0
                     break
-            elif isinstance(condVal, list):
+            elif isinstance(condVal, list) and isinstance(condVal[0], Number):
                 if self.tags.get(condKey) < condVal[0] or self.tags.get(condKey) > condVal[1]:
                     conditionsMet = 0
                     break
+            elif isinstance(condVal, list) and isinstance(condVal[0], str):
+                if self.tags[condKey] not in condVal:
+                    conditionsMet = 0
+                    break 
             elif self.tags[condKey] != condVal: 
                 conditionsMet = 0
                 break
@@ -75,16 +81,16 @@ class Cell (object):
         for sectName,sectParams in prop['secs'].iteritems(): 
             # create section
             if sectName not in self.secs:
-                self.secs[sectName] = {}  # create section dict
+                self.secs[sectName] = Dict()  # create section dict
             sec = self.secs[sectName]  # pointer to section
             
             # add distributed mechanisms 
             if 'mechs' in sectParams:
                 for mechName,mechParams in sectParams['mechs'].iteritems(): 
                     if 'mechs' not in sec:
-                        sec['mechs'] = {}
+                        sec['mechs'] = Dict()
                     if mechName not in sec['mechs']: 
-                        sec['mechs'][mechName] = {}  
+                        sec['mechs'][mechName] = Dict()  
                     for mechParamName,mechParamValue in mechParams.iteritems():  # add params of the mechanism
                         sec['mechs'][mechName][mechParamName] = mechParamValue
 
@@ -93,9 +99,9 @@ class Cell (object):
                 for pointpName,pointpParams in sectParams['pointps'].iteritems(): 
                     #if self.tags['cellModel'] == pointpName: # only required if want to allow setting various cell models in same rule
                     if 'pointps' not in sec:
-                        sec['pointps'] = {}
+                        sec['pointps'] = Dict()
                     if pointpName not in sec['pointps']: 
-                        sec['pointps'][pointpName] = {}  
+                        sec['pointps'][pointpName] = Dict()  
                     for pointpParamName,pointpParamValue in pointpParams.iteritems():  # add params of the mechanism
                         sec['pointps'][pointpName][pointpParamName] = pointpParamValue
 
@@ -104,7 +110,7 @@ class Cell (object):
             if 'geom' in sectParams:
                 for geomParamName,geomParamValue in sectParams['geom'].iteritems():  
                     if 'geom' not in sec:
-                        sec['geom'] = {}
+                        sec['geom'] = Dict()
                     if not type(geomParamValue) in [list, dict]:  # skip any list or dic params
                         sec['geom'][geomParamName] = geomParamValue
 
@@ -118,7 +124,7 @@ class Cell (object):
             # add topolopgy params
             if 'topol' in sectParams:
                 if 'topol' not in sec:
-                    sec['topol'] = {}
+                    sec['topol'] = Dict()
                 for topolParamName,topolParamValue in sectParams['topol'].iteritems(): 
                     sec['topol'][topolParamName] = topolParamValue
 
@@ -134,31 +140,30 @@ class Cell (object):
             self.secLists.update(prop['secLists'])  # diction of section lists
 
 
-
     def initV (self): 
         for sec in self.secs.values():
             if 'vinit' in sec:
-                sec['hSection'].v = sec['vinit']
+                sec['hSec'].v = sec['vinit']
 
     def createNEURONObj (self, prop):
         # set params for all sections
         for sectName,sectParams in prop['secs'].iteritems(): 
             # create section
             if sectName not in self.secs:
-                self.secs[sectName] = {}  # create sect dict if doesn't exist
-            if not self.secs[sectName].get('hSection'): 
-                self.secs[sectName]['hSection'] = h.Section(name=sectName)  # create h Section object
+                self.secs[sectName] = Dict()  # create sect dict if doesn't exist
+            if not self.secs[sectName].get('hSec'): 
+                self.secs[sectName]['hSec'] = h.Section(name=sectName)  # create h Section object
             sec = self.secs[sectName]  # pointer to section
             
             # add distributed mechanisms 
             if 'mechs' in sectParams:
-                for mechName,mechParams in sectParams['mechs'].iteritems():  
+                for mechName,mechParams in sectParams['mechs'].iteritems(): 
                     if mechName not in sec['mechs']: 
-                        sec['mechs'][mechName] = {}
-                    sec['hSection'].insert(mechName)
+                        sec['mechs'][mechName] = Dict()
+                    sec['hSec'].insert(mechName)
                     for mechParamName,mechParamValue in mechParams.iteritems():  # add params of the mechanism
                         mechParamValueFinal = mechParamValue
-                        for iseg,seg in enumerate(sec['hSection']):  # set mech params for each segment
+                        for iseg,seg in enumerate(sec['hSec']):  # set mech params for each segment
                             if type(mechParamValue) in [list]: 
                                 mechParamValueFinal = mechParamValue[iseg]
                             seg.__getattribute__(mechName).__setattr__(mechParamName,mechParamValueFinal)
@@ -174,10 +179,10 @@ class Cell (object):
                 for pointpName,pointpParams in sectParams['pointps'].iteritems(): 
                     #if self.tags['cellModel'] == pointpParams:  # only required if want to allow setting various cell models in same rule
                     if pointpName not in sec['pointps']:
-                        sec['pointps'][pointpName] = {} 
+                        sec['pointps'][pointpName] = Dict() 
                     pointpObj = getattr(h, pointpParams['mod'])
                     loc = pointpParams['loc'] if 'loc' in pointpParams else 0.5  # set location
-                    sec['pointps'][pointpName]['hPointp'] = pointpObj(loc, sec = sec['hSection'])  # create h Pointp object (eg. h.Izhi2007b)
+                    sec['pointps'][pointpName]['hPointp'] = pointpObj(loc, sec = sec['hSec'])  # create h Pointp object (eg. h.Izhi2007b)
                     for pointpParamName,pointpParamValue in pointpParams.iteritems():  # add params of the point process
                         if pointpParamName not in ['mod', 'loc', 'vref', 'synList'] and not pointpParamName.startswith('_'):
                             setattr(sec['pointps'][pointpName]['hPointp'], pointpParamName, pointpParamValue)
@@ -186,11 +191,11 @@ class Cell (object):
             if 'geom' in sectParams:
                 for geomParamName,geomParamValue in sectParams['geom'].iteritems():  
                     if not type(geomParamValue) in [list, dict]:  # skip any list or dic params
-                        setattr(sec['hSection'], geomParamName, geomParamValue)
+                        setattr(sec['hSec'], geomParamName, geomParamValue)
 
                 # set 3d geometry
                 if 'pt3d' in sectParams['geom']:  
-                    h.pt3dclear(sec=sec['hSection'])
+                    h.pt3dclear(sec=sec['hSec'])
                     x = self.tags['x']
                     if 'ynorm' in self.tags and hasattr(sim.net.params, 'sizeY'):
                         y = self.tags['ynorm'] * sim.net.params.sizeY/1e3  # y as a func of ynorm and cortical thickness
@@ -198,14 +203,14 @@ class Cell (object):
                         y = self.tags['y']
                     z = self.tags['z']
                     for pt3d in sectParams['geom']['pt3d']:
-                        h.pt3dadd(x+pt3d[0], y+pt3d[1], z+pt3d[2], pt3d[3], sec=sec['hSection'])
+                        h.pt3dadd(x+pt3d[0], y+pt3d[1], z+pt3d[2], pt3d[3], sec=sec['hSec'])
 
         # set topology 
         for sectName,sectParams in prop['secs'].iteritems():  # iterate sects again for topology (ensures all exist)
             sec = self.secs[sectName]  # pointer to section # pointer to child sec
             if 'topol' in sectParams:
                 if sectParams['topol']:
-                    sec['hSection'].connect(self.secs[sectParams['topol']['parentSec']]['hSection'], sectParams['topol']['parentX'], sectParams['topol']['childX'])  # make topol connection
+                    sec['hSec'].connect(self.secs[sectParams['topol']['parentSec']]['hSec'], sectParams['topol']['parentX'], sectParams['topol']['childX'])  # make topol connection
 
 
     # Create NEURON objs for conns and syns if included in prop (used when loading)
@@ -216,8 +221,8 @@ class Cell (object):
                 self.addNetStim (stimParams, stimContainer=stimParams)
        
             elif stimParams['type'] in ['IClamp', 'VClamp', 'SEClamp', 'AlphaSynapse']:
-                stim = getattr(h, stimParams['type'])(self.secs[stimParams['sec']]['hSection'](stimParams['loc']))
-                stimProps = {k:v for k,v in stimParams.iteritems() if k not in ['type', 'source', 'loc', 'sec', 'h'+stimParams['type']]}
+                stim = getattr(h, stimParams['type'])(self.secs[stimParams['sec']]['hSec'](stimParams['loc']))
+                stimProps = {k:v for k,v in stimParams.iteritems() if k not in ['label', 'type', 'source', 'loc', 'sec', 'h'+stimParams['type']]}
                 for stimPropName, stimPropValue in stimProps.iteritems(): # set mechanism internal stimParams
                     if isinstance(stimPropValue, list):
                         if stimPropName == 'amp': 
@@ -277,10 +282,10 @@ class Cell (object):
             if 'pointps' in sec:  # if no syns, check if point processes with 'vref' (artificial cell)
                 for pointpName, pointpParams in sec['pointps'].iteritems():
                     if 'vref' in pointpParams:
-                        nc = h.NetCon(sec['pointps'][pointpName]['hPointp'].__getattribute__('_ref_'+pointpParams['vref']), None, sec=sec['hSection'])
+                        nc = h.NetCon(sec['pointps'][pointpName]['hPointp'].__getattribute__('_ref_'+pointpParams['vref']), None, sec=sec['hSec'])
                         break
             if not nc:  # if still haven't created netcon  
-                nc = h.NetCon(sec['hSection'](loc)._ref_v, None, sec=sec['hSection'])
+                nc = h.NetCon(sec['hSec'](loc)._ref_v, None, sec=sec['hSec'])
             nc.threshold = threshold
             sim.pc.cell(self.gid, nc, 1)  # associate a particular output stream of events
             sim.net.gid2lid[self.gid] = len(sim.net.lid2gid)
@@ -298,7 +303,7 @@ class Cell (object):
                     sec['synMechs'] = []
                 synMech = next((synMech for synMech in sec['synMechs'] if synMech['label']==synLabel and synMech['loc']==loc), None)
                 if not synMech:  # if synMech not in section, then create
-                    synMech = {'label': synLabel, 'loc': loc}
+                    synMech = Dict({'label': synLabel, 'loc': loc})
                     sec['synMechs'].append(synMech)
 
             if sim.cfg.createNEURONObj:
@@ -308,11 +313,11 @@ class Cell (object):
                 if not synMech:  # if pointer not created in createPyStruct, then check 
                     synMech = next((synMech for synMech in sec['synMechs'] if synMech['label']==synLabel and synMech['loc']==loc), None)
                 if not synMech:  # if still doesnt exist, then create
-                    synMech = {}
+                    synMech = Dict()
                     sec['synMechs'].append(synMech)
                 if not synMech.get('hSyn'):  # if synMech doesn't have NEURON obj, then create
                     synObj = getattr(h, synMechParams['mod'])
-                    synMech['hSyn'] = synObj(loc, sec = sec['hSection'])  # create h Syn object (eg. h.Exp2Syn)
+                    synMech['hSyn'] = synObj(loc, sec = sec['hSec'])  # create h Syn object (eg. h.Exp2Syn)
                     for synParamName,synParamValue in synMechParams.iteritems():  # add params of the synaptic mechanism
                         if synParamName not in ['label', 'mod']:
                             setattr(synMech['hSyn'], synParamName, synParamValue)
@@ -368,10 +373,10 @@ class Cell (object):
                 if netStimParams:
                     connParams['preGid'] = 'NetStim'
                     connParams['preLabel'] = netStimParams['source']
-                self.conns.append(connParams)
+                self.conns.append(Dict(connParams))
                     
             else:  # do not fill in python structure (just empty dict for NEURON obj)
-                self.conns.append({})
+                self.conns.append(Dict())
 
             # NEURON objects
             if sim.cfg.createNEURONObj:
@@ -417,10 +422,14 @@ class Cell (object):
                         compareTo = conn.get(condKey)
 
                     # check if conditions met
-                    if isinstance(condVal, list):
+                    if isinstance(condVal, list) and isinstance(condVal[0], Number):
                         if compareTo < condVal[0] or compareTo > condVal[1]:
                             conditionsMet = 0
                             break
+                    elif isinstance(condVal, list) and isinstance(condVal[0], str):
+                        if self.tags[condKey] not in condVal:
+                            conditionsMet = 0
+                            break 
                     elif compareTo != condVal: 
                         conditionsMet = 0
                         break
@@ -428,10 +437,14 @@ class Cell (object):
             if conditionsMet and 'postConds' in params:
                 for (condKey,condVal) in params['postConds'].iteritems():  # check if all conditions are met
                     # check if conditions met
-                    if isinstance(condVal, list):
+                    if isinstance(condVal, list) and isinstance(condVal[0], Number):
                         if self.tags.get(condKey) < condVal[0] or self.tags.get(condKey) > condVal[1]:
                             conditionsMet = 0
                             break
+                    elif isinstance(condVal, list) and isinstance(condVal[0], str):
+                        if self.tags[condKey] not in condVal:
+                            conditionsMet = 0
+                            break 
                     elif self.tags.get(condKey) != condVal: 
                         conditionsMet = 0
                         break
@@ -460,10 +473,14 @@ class Cell (object):
             
             for (condKey,condVal) in params['conds'].iteritems():  # check if all conditions are met
                 # check if conditions met
-                if isinstance(condVal, list):
+                if isinstance(condVal, list) and isinstance(condVal[0], Number):
                     if stim.get(condKey) < condVal[0] or stim.get(condKey) > condVal[1]:
                         conditionsMet = 0
                         break
+                elif isinstance(condVal, list) and isinstance(condVal[0], str):
+                    if self.tags[condKey] not in condVal:
+                        conditionsMet = 0
+                        break 
                 elif stim[condKey] != stim[condKey]: 
                     conditionsMet = 0
                     break
@@ -504,7 +521,7 @@ class Cell (object):
 
     def addNetStim (self, params, stimContainer=None):
         if not stimContainer:
-            self.stims.append(params.copy())  # add new stim to Cell object
+            self.stims.append(Dict(params.copy()))  # add new stim to Cell object
             stimContainer = self.stims[-1]
         
         rand = h.Random()
@@ -575,7 +592,7 @@ class Cell (object):
        
 
         elif params['type'] in ['IClamp', 'VClamp', 'SEClamp', 'AlphaSynapse']:
-            stim = getattr(h, params['type'])(sec['hSection'](params['loc']))
+            stim = getattr(h, params['type'])(sec['hSec'](params['loc']))
             stimParams = {k:v for k,v in params.iteritems() if k not in ['type', 'source', 'loc', 'sec', 'label']}
             stringParams = ''
             for stimParamName, stimParamValue in stimParams.iteritems(): # set mechanism internal params
@@ -590,7 +607,7 @@ class Cell (object):
                 else: 
                     setattr(stim, stimParamName, stimParamValue)
                     stringParams = stringParams + ', ' + stimParamName +'='+ str(stimParamValue)
-            self.stims.append(params) # add to python structure
+            self.stims.append(Dict(params)) # add to python structure
             self.stims[-1]['h'+params['type']] = stim  # add stim object to dict in stims list
 
             if sim.cfg.verbose: print('  Added %s %s to cell gid=%d, sec=%s, loc=%.4g%s'%
@@ -720,7 +737,7 @@ class Cell (object):
 
     def _distributeSynsUniformly (self, secList, numSyns):
         from numpy import cumsum
-        secLengths = [self.secs[s]['hSection'].L for s in secList]
+        secLengths = [self.secs[s]['hSec'].L for s in secList]
         totLength = sum(secLengths)
         cumLengths = list(cumsum(secLengths))
         absLocs = [i*(totLength/numSyns)+totLength/numSyns/2 for i in range(numSyns)]
@@ -734,7 +751,7 @@ class Cell (object):
         plasticity = params.get('plasticity')
         if plasticity and sim.cfg.createNEURONObj:
             try:
-                plastMech = getattr(h, plasticity['mech'], None)(0, sec=sec['hSection'])  # create plasticity mechanism (eg. h.STDP)
+                plastMech = getattr(h, plasticity['mech'], None)(0, sec=sec['hSec'])  # create plasticity mechanism (eg. h.STDP)
                 for plastParamName,plastParamValue in plasticity['params'].iteritems():  # add params of the plasticity mechanism
                     setattr(plastMech, plastParamName, plastParamValue)
                 if plasticity['mech'] == 'STDP':  # specific implementation steps required for the STDP mech
@@ -758,13 +775,13 @@ class Cell (object):
                 ptr = None
                 if 'loc' in params:
                     if 'mech' in params:  # eg. soma(0.5).hh._ref_gna
-                        ptr = self.secs[params['sec']]['hSection'](params['loc']).__getattribute__(params['mech']).__getattribute__('_ref_'+params['var'])
+                        ptr = self.secs[params['sec']]['hSec'](params['loc']).__getattribute__(params['mech']).__getattribute__('_ref_'+params['var'])
                     elif 'synMech' in params:  # eg. soma(0.5).AMPA._ref_g
                         sec = self.secs[params['sec']]
                         synMech = next((synMech for synMech in sec['synMechs'] if synMech['label']==params['synMech'] and synMech['loc']==params['loc']), None)
                         ptr = synMech['hSyn'].__getattribute__('_ref_'+params['var'])
                     else:  # eg. soma(0.5)._ref_v
-                        ptr = self.secs[params['sec']]['hSection'](params['loc']).__getattribute__('_ref_'+params['var'])
+                        ptr = self.secs[params['sec']]['hSec'](params['loc']).__getattribute__('_ref_'+params['var'])
                 else:
                     if 'pointp' in params: # eg. soma.izh._ref_u
                         if params['pointp'] in self.secs[params['sec']]['pointps']:
@@ -779,7 +796,7 @@ class Cell (object):
 
 
     def recordStimSpikes (self):
-        sim.simData['stims'].update({'cell_'+str(self.gid): {}})
+        sim.simData['stims'].update({'cell_'+str(self.gid): Dict()})
         for conn in self.conns:
             if conn['preGid'] == 'NetStim':
                 stimSpikeVecs = h.Vector() # initialize vector to store