+++ CHANGES.md
@@ -1,3 +1,17 @@
+## Version 0.5.4
+
+- Access to nested dictionaries via dot notation (secs.soma.geom) using new inherited classes Dict() and ODict() (issue #123)
+
+- 'hSection' replaced with 'hSec' 
+
+- Added sim.popAvgRates() to calculate and optionally print population avg rates 
+
+- Functions to modify cells, conns or stims also update sim.allCells (issue #59)
+
+- Fixed bug so can modify funcs now accept lists of strings (issue #59)
+
+- Fixed bug when recording traces of relative cell indices of populations across multiple nodes
+
 ## Version 0.5.3
 
 - Functions to modify parameters of cells, conns and stims in an instantiated network (issue #59)